requirements.txt
Django==3.1.4
djangorestframework